Add support for the PHP 5.4 builtin web server
PHP version 5.4 brought with it a new built-in command line web server
that can be run using the `php -S localhost:4000 index.php` command.
The [Silex documentation][1] has more details, and, while the server is
not intended for production use, it's a huge time saver for local
development.

Unfortunately, Bolt doesn't currently work with this feature for four
reasons:

1. The logic in the `getPaths` function breaks asset paths when used
with the built-in server

2. The index.php understandably doesn't yield to static assets as the
server expects it to.

3. The CLI server doesn't understand the subtleties of the `thumbs`
route.

4. The `timthumb` class references images and file paths in a relative
manner, making it very fragile when the entry point changes. Perhaps a
better route would be to make these path'd absolutely using some of the
directory constants.

For all of these problems a comparison with `php_sapi_name()` and
'cli-server' is performed. If the app is running with the CLI server, it
returns false when a static asset is called. If the asset is via the
`thumbs` URL then the server performs the necessary manipulations to
send the request to the right place.

Note: I chose not to pollute the global constant namespace and continue
to check `php_sapi_name()` in every instance, although I doubt there's a
significant performance penalty.

Finally, while I did directly compare the cli-server with a correctly
configured Nginx instance, I haven't done any testing against Apache or
others. This commit also doesn't include any test cases.