Okay so I got this to work!
.... but this doesn't make it pretty!
We'll provide, just like imagelists, a placeholder image along with a
changeable description. Another thing to, maybe, provide might be a
filesize. But as the files are static on the server laying around as
files, I think their output may be read from another place and should'nt
lay around within the DB.