Feature: Additional Dropdown in Mainmenu for custom Contenttypes
With this feature you have the possibility to add a Dropdown in the
Mainmenu for some Contenttypes. To activate set „showinmenu = true“ for
needed contenttypes in contenttypes settings.

Contenttypes with this additional Dropdown are automatically hidden in
the default Content Dropownmenu.

The idea behind this feature is to allow Quick access for related
contents. Enjoy!