Arbitrary Fix for #965
As the container gets its value set to the stuff that's being returned
by the service (itself: the pasted content), the complete content of the
markdown textarea gets deleted and overwritten by the pasted stuff.
Yikes. This is NOT the final solution, as we need to find out where
within the textarea something was actually pasted. This could become
intresting, since references are supposed to go at the bottom of the
document. Meh.