Fixed a silly bug, related to #1070.
The "smart" unserializer used to be uncritically applied to all field
values; this caused text field contents that happened to be valid
serialized arrays to be displayed as just "Array" rather than the actual
text. With JSON decoding added to the smart unserializer, the same
happened with valid JSON arrays and objects, so entering something like
`[]` or `{}` or `["foo", "bar"]` in a textarea would trigger JSON
decoding.

The fix is to check the field type first, and only attempt decoding on
field types that expect it (image, file, video, etc., but not text,
markdown, html and such).