Add asynchronous social media library based on Socialite https://github.com/tmort/Socialite
The over all effect is faster intial pages loads as buttons and widgets
rendered are done in JS after page load has finished and the user
scrolls to the area of the page.

Code has been baking for a month in my staging environment and passes
Google PageSpeed test nicely!