Fixed XSS in snippets generator
Under certain circumstances, parts of the client-supplied URL path would
be used in the "canonical" <link>, and no HTML-encoding was applied to
secure this, causing a reflected XSS vulnerability.

I have not been able to reproduce the exploit on my test environment,
but I have verified it on another system.