Fixed a fatal error when adding order=... to query string.
When adding an order=... parameter to the URL of any request that called
getContent() without an explicit ordering, Bolt would uncritically pass
the requested ordering through to the SQL backend. For MySQL, this would
cause a fatal error when the requested column didn't exist. The fact
that the parameter got passed through pretty much unchecked also means
that there was a security problem there; I haven't found a way to
exploit it though.

Either way, I removed this "feature", and as far as I can tell, nothing
breaks.