Fix testDbCheck.
Config::checkConfig calls IntegrityChecker::needsCheck which returns false when the dbcheck_ts doesn't exist.
Since it does need check, checkTablesIntegrity is called.
This counts as the one invocation call testDbCheck is expecting.
The call from Backend::dbCheck counts as the second invocation - failing the test.

Since the test is for the dbCheck route we don't care what Config::checkConfig is doing. That should be tested elsewhere.