New config setting additional_locales
I found that datelocal was not actually using the locale that I had
configured. This seems to be caused by environment specific issues. A
way to solve it is by specifying additional locales. This change adds an
optional config parameter where additional locales can be configured. If
configured they are appended to the local array. Using this, adding
"Dutch_Netherlands" I could get Dutch dates (which did not work by
simply setting nl_NL as the locale)