Add IIS web.config
Add in IIS URL Rewrite module drop in replacement for Apache .htaccess
file. URL rewriting is handled, as is the woff2 file definition, but I
had to disable the cache access and yml,db,twig,md deny rules as they
interfered with the CMS backend.

WARNING: In order to have this work on IIS you must have the IIS URL
Rewrite module. You will also need to alter the permissions on the files
folder and the PHP extension php_fileinfo will need to be enabled.
Instructions to follow in the documentation repository.