Add # BEGIN/END CONTENTTYPE NAME
I started doing this in my own templates to quickly recognize the contenttype separations with 'big' contenttypes. 

Admittedly, there's not much added value when scrolling down (you will see the contenttype name in the first line anyway) but it helps to know where you are when scrolling up.

See if you like the idea. If so, it could be added as a Geheimtip to the Docs, too.