Increase MySQL group_concat_max_len variable
I've been seeing some issues with the repeater content type; in essence,
adding more than 10 or so bits of content seems to cause it to fail to
save properly – repeater sections end up missing.

@rossriley did some investigation after I brought this up on Slack
(thanks Ross!) and reached the conclusion that this was due to the
result of a GROUP_CONCAT() operation exceeding 1024 bytes in MySQL. When
this happens, by default MySQL silently starts chopping off extra bytes.
Good old MySQL.

This adds an extra query to the `postConnect()` callback of
`EventListener\DoctrineListener` which increases the GROUP_CONCAT()
result limit to 100,000 bytes if we're connected to a MySQL server,
which should mitigate this from happening.

There could be a better (and somewhat less hacky) way of stopping this
happening, however this appears to be the most reliable with minimal
code change.