Removed BC check for editing file, because it was invalid. Since no one has complained about it since Feb 2015 I think we are good to remove.
The old url that we were providing BC for was "/bolt/files/edit/app/config/config.yml". It was supposed to be converted to config://config.yml,
but it was actually producing config://config/config.yml. That file didn't exist and we redirected to dashboard. I think all the links in templates
were converted when this error should up awhile ago, which was the right thing to do instead of fixing this hack.