Minor feature: singletons
This checks if there is only one record and the current user has no
permissions to create new records and will then consider that
contenttype a "singleton" that we link directly to the only record,
skipping the submenu. Also updates the tests to expect showcases to be a
singleton in the editor tests.