Remove default value for cookies_domain. Fixes #6431
The only thing this did differently than PHP's default domain is removed
the www. prefix allowing the cookie to assigned to foo.com and www.foo.com.

It was decided this isn't a big enough use case to warrant keeping this logic.
Note that fixing this caching bug would mean moving the logic to 3 different
places (or 2 with larger refactoring).