Replaced `display_errors` and `reporting_level` with `throw_at` and `log_at` DI parameters.
ErrorHandler's "displayErrors" logic is deprecated and we are instructed to use "throw at" logic instead.
This is also better because instead of a boolean we can specify which errors we want thrown (aka displayed to user).

Additionally, PHP has `display_errors` ini setting which is similar but not the same as ErrorHandler's "display errors" logic.
This is now automatically disabled (if error handler is enabled) since it is not helpful to users.
You can see the error with the logger you've configured and optionally see it inline as an exception message.

PHP's `reporting_level` is automatically configured to report all errors (if error handler is enabled).
This is because Symfony's ErrorHandler has its own logic for configuring how errors are handled: "throw at" and "log_at".