Add roave/security-advisories to require-dev
This package can only be required in its dev-master version: there will
never be stable/tagged versions because of the nature of the problem
being targeted. Security issues are in fact a moving target, and locking
your project to a specific tagged version of the package would not make
any sense.