Added method to login as specific user without pasword
As discussed on Slack, the added `loinAsUser()` method enabled extension creators to override the login mechanics and connect bolt to various kinds of SSO systems. Since the logic for user management and authentication is currently (pre 4.0) not centralized enough, a lot of needed code would have to be duplicated to archive this. (And we certainly dont want to go there ;) )

PS: The file is identical to 3.5 and could also be backported?

I can provide guides on how to do this if needed.